% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{deleteJob}
\alias{deleteJob}
\title{Delete a job}
\usage{
deleteJob(jobId, deleteResult = TRUE)
}
\arguments{
\item{jobId}{A job id}

\item{deleteResult}{TRUE to delete job result in storage blob,
FALSE to keep the result in storage blob}
}
\description{
Delete a job
}
\examples{
\dontrun{
deleteJob("job-001")
deleteJob("job-001", deleteResult = FALSE)
}
}
