% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_management.R
\name{deleteStorageContainer}
\alias{deleteStorageContainer}
\title{Delete a storage container from Azure Storage}
\usage{
deleteStorageContainer(container)
}
\arguments{
\item{container}{The name of the container}
}
\description{
Delete a storage container from Azure Storage
}
\examples{
deleteStorageContainer("outputContainerName")
}
