% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage-api.R
\name{deleteStorageFile}
\alias{deleteStorageFile}
\title{Delete a storage file from a container.}
\usage{
deleteStorageFile(container, blobPath, ...)
}
\arguments{
\item{container}{The name of container}

\item{blobPath}{The file path of the blob}
}
\description{
Delete a storage file from a container.
}
