% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{generateClusterConfig}
\alias{generateClusterConfig}
\title{Creates a configuration file for the user's cluster setup.}
\usage{
generateClusterConfig(fileName, ...)
}
\arguments{
\item{fileName}{A string that uniquely identifies the job within the account.}

\item{...}{Further named parameters
\itemize{
 \item{"batchAccount"}: {A list of files that the Batch service will download to the compute node before running the command line.}
 \item{"batchKey"}: {Arguments in the foreach parameters that will be used for the task running.}
 \item{"batchUrl"}: {A list of packages that the Batch service will download to the compute node.}
 \item{"storageAccount"}: {The R environment that the task will run under.}
 \item{"storageKey"}: {The R environment that the task will run under.}
}}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates a configuration file for the user's cluster setup.
}
\examples{
{
generateClusterConfig("test_config.json")
generateClusterConfig("test_config.json", batchAccount = "testbatchaccount", batchKey = "test_batch_account_key", batchUrl = "http://testbatchaccount.azure.com", storageAccount = "teststorageaccount", storageKey = "test_storage_account_key")
}
}

