% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{generateCredentialsConfig}
\alias{generateCredentialsConfig}
\title{Creates a credentials file for rAzureBatch package authentication}
\usage{
generateCredentialsConfig(fileName, authenticationType = "SharedKey", ...)
}
\arguments{
\item{fileName}{Credentials file name}

\item{authenticationType}{The type of authentication for Azure: SharedKey, ServicePrincipal}

\item{...}{Further named parameters
\itemize{
 \item{"batchAccount"}: {Batch account name for Batch Service authentication.}
 \item{"batchKey"}: {Batch account key for signing REST signatures.}
 \item{"batchUrl"}: {Batch service url for account.}
 \item{"storageAccount"}: {Storage account for storing output results.}
 \item{"storageKey"}: {Storage account key for storage service authentication.}
 \item{"storageEndpointSuffix"}: {Values: core.windows.net, core.chinacloudapi.cn, core.cloudapi.de, core.usgovcloudapi.net }
 \item{"githubAuthenticationToken"}: {GitHub authentication token for pulling R
                                      packages from private GitHub repositories}
 \item{"dockerAuthentication"}: {Docker authentication for pulling Docker images
                                 from private Docker registries}
 \item{"dockerUsername"}: {Username to docker registry}
 \item{"dockerPassword"}: {Password to docker registry}
 \item{"dockerRegistry"}: {URL to docker registry}

}}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates a credentials file for rAzureBatch package authentication
}
\examples{
{
generateCredentialsConfig("test_config.json")
generateCredentialsConfig("test_config.json", batchAccount = "testbatchaccount",
   batchKey = "test_batch_account_key", batchUrl = "http://testbatchaccount.azure.com",
   storageAccount = "teststorageaccount", storageKey = "test_storage_account_key",
   storageEndpointSuffix = "core.windows.net")
}
}
