% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{getCluster}
\alias{getCluster}
\title{Gets the cluster from your Azure account.}
\usage{
getCluster(clusterName)
}
\arguments{
\item{clusterName}{The cluster configuration that was created in \code{makeCluster}}
}
\description{
Gets the cluster from your Azure account.
}
\examples{
\dontrun{
cluster <- getCluster("myCluster")
}
}
