% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{getClusterFile}
\alias{getClusterFile}
\title{Get node files from compute nodes. By default, this operation will print the files on screen.}
\usage{
getClusterFile(cluster, nodeId, filePath, verbose = TRUE, overwrite = FALSE,
  downloadPath = NULL)
}
\arguments{
\item{cluster}{The cluster object}

\item{nodeId}{Id of the node}

\item{filePath}{The path to the file that you want to get the contents of}

\item{verbose}{Flag for printing log files onto console}

\item{...}{Further named parameters
\itemize{
 \item{"downloadPath"}: { Path to save file to }
 \item{"overwrite"}: { Will only overwrite existing localPath }
}}
}
\description{
Get node files from compute nodes. By default, this operation will print the files on screen.
}
\examples{
\dontrun{
stdoutText <- getClusterFile(cluster, "tvm-1170471534_1-20170829t072146z",
filePath = "stdout.txt", verbose = FALSE)
getClusterFile(cluster, "tvm-1170471534_2-20170829t072146z",
filePath = "wd/output.csv", downloadPath = "output.csv", overwrite = TRUE)
}
}
