% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{getClusterList}
\alias{getClusterList}
\title{Get a list of clusters by state from the given filter}
\usage{
getClusterList(filter = NULL)
}
\arguments{
\item{filter}{A filter containing cluster state}
}
\description{
Get a list of clusters by state from the given filter
}
\examples{
\dontrun{
getClusterList()
}
}
