% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{getClusterLogs}
\alias{getClusterLogs}
\title{The registerDoAzureParallel function is used to register
the Azure cloud-enabled parallel backend with the foreach package.}
\usage{
getClusterLogs(cluster, nodeId, type = c("startup", "stdout", "stderr"), ...)
}
\arguments{
\item{cluster}{The cluster object}

\item{nodeId}{Id of the node}

\item{type}{The type of logs the user wants}

\item{...}{Further named parameters
\itemize{
 \item{"write"}: {Batch account name for Batch Service authentication.}
}}
}
\description{
The registerDoAzureParallel function is used to register
the Azure cloud-enabled parallel backend with the foreach package.
}
\examples{
getClusterLogs(cluster)
}
