% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobUtilities.R
\name{getJob}
\alias{getJob}
\title{Get a job for the given job id}
\usage{
getJob(jobId, verbose = TRUE)
}
\arguments{
\item{jobId}{A job id}

\item{verbose}{show verbose log output}
}
\description{
Get a job for the given job id
}
\examples{
\dontrun{
getJob("job-001", FALSE)
}
}
