% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-operations.R
\name{getJobFile}
\alias{getJobFile}
\title{Get job-related files from cluster node. By default, this operation will print the files on screen.}
\usage{
getJobFile(jobId, taskId, filePath, downloadPath = NULL, verbose = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{jobId}{Id of the foreach job}

\item{taskId}{Id of the task}

\item{filePath}{the path to the task file that you want to get the contents of}

\item{verbose}{Flag for printing the log files onto console}

\item{...}{Further named parameters
\itemize{
 \item{"downloadPath"}: { Path to save file to }
 \item{"overwrite"}: { Will only overwrite existing localPath }
}}
}
\description{
Get job-related files from cluster node. By default, this operation will print the files on screen.
}
\examples{
\dontrun{
stdoutFile <- getJobFile("job20170822055031", "job20170822055031-task1", "stderr.txt")
getJobFile("job20170822055031", "job20170822055031-task1", "stdout.txt", downloadPath = "hello.txt")
}
}
