% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getJobList}
\alias{getJobList}
\title{Get a list of job statuses from the given filter}
\usage{
getJobList(filter = NULL)
}
\arguments{
\item{filter}{A filter containing job state}
}
\description{
Get a list of job statuses from the given filter
}
\examples{
\dontrun{
getJobList()
}
}
