% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{getJobLogs}
\alias{getJobLogs}
\title{Enables}
\usage{
getJobLogs(jobId, taskId, type = c("rlogs", "stdout", "stderr"), ...)
}
\arguments{
\item{jobId}{The cluster object to use for parallelization}

\item{type}{}
}
\description{
Enables
}
\examples{
getJobLogs(cluster)
}
