% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_management.R
\name{getStorageFile}
\alias{getStorageFile}
\title{Get a storage file from Azure Storage. By default, this operation will print the files on screen.}
\usage{
getStorageFile(container, blobPath, downloadPath = NULL, overwrite = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{container}{The name of the container}

\item{blobPath}{The path of the blob}

\item{...}{Optional parameters
\itemize{
 \item{"downloadPath"}: { Path to save file to }
 \item{"overwrite"}: { Will only overwrite existing localPath }
 \item{"verbose"}: { Show verbose messages }
}}
}
\description{
Get a storage file from Azure Storage. By default, this operation will print the files on screen.
}
\examples{
\dontrun{
stdoutText <- getStorageFile(testContainer, "logs/stdout.txt")
}
}
