% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_management.R
\name{listStorageContainers}
\alias{listStorageContainers}
\title{List storage containers from Azure Storage.}
\usage{
listStorageContainers(prefix = "")
}
\arguments{
\item{prefix}{Filters the results to return only containers
whose name begins with the specified prefix.}
}
\description{
List storage containers from Azure Storage.
}
\examples{
containers <- listStorageContainers()
View(containers)
}
