% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_management.R
\name{listStorageFiles}
\alias{listStorageFiles}
\title{List storage files from Azure storage.}
\usage{
listStorageFiles(container, prefix = "", ...)
}
\arguments{
\item{container}{The cluster object}

\item{prefix}{Id of the node}
}
\description{
List storage files from Azure storage.
}
\examples{
files <- listStorageFiles("job001")
View(files)
}
