% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{makeCluster}
\alias{makeCluster}
\title{Creates an Azure cloud-enabled cluster.}
\usage{
makeCluster(clusterSetting = "cluster_settings.json", fullName = FALSE,
  wait = TRUE, resourceFiles = list())
}
\arguments{
\item{clusterSetting}{Cluster configuration object or file name}

\item{fullName}{A boolean flag for checking the file full name}

\item{wait}{A boolean flag to wait for all nodes to boot up}

\item{resourceFiles}{A list of files that Batch will download to the compute node before running the command line}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates an Azure cloud-enabled cluster.
}
\examples{
\dontrun{
cluster <- makeCluster("cluster_config.json", fullName = TRUE, wait = TRUE)
}
}
