% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{makeClusterObject}
\alias{makeClusterObject}
\title{Creates an Azure cloud-enabled cluster.}
\usage{
makeClusterObject(clusterObject, wait = TRUE, resourceFiles = list())
}
\arguments{
\item{wait}{A boolean flag to wait for all nodes to boot up}

\item{resourceFiles}{A list of files that Batch will download to the compute node before running the command line}

\item{clusterSettingsObj}{Cluster configuration as R object}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates an Azure cloud-enabled cluster.
}
\examples{
\dontrun{
cluster <- makeClusterObject(clusterSettingsObj = clusterConfigObj, fullName = TRUE, wait = TRUE)
}
}
