% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{registerDoAzureParallel}
\alias{registerDoAzureParallel}
\title{The registerDoAzureParallel function is used to register the Azure cloud-enabled parallel backend with the foreach package.}
\usage{
registerDoAzureParallel(cluster)
}
\arguments{
\item{cluster}{The cluster object to use for parallelization}
}
\description{
The registerDoAzureParallel function is used to register the Azure cloud-enabled parallel backend with the foreach package.
}
\examples{
registerDoAzureParallel(cluster)
}
