% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{setAutoDeleteJob}
\alias{setAutoDeleteJob}
\title{Specify whether to delete job and its result after asychronous job is completed.}
\usage{
setAutoDeleteJob(value = TRUE)
}
\arguments{
\item{value}{boolean of TRUE or FALSE}
}
\description{
Specify whether to delete job and its result after asychronous job is completed.
}
\examples{
setAutoDeleteJob(FALSE)
}
