% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{setChunkSize}
\alias{setChunkSize}
\title{Groups iterations of the foreach loop together per task.}
\usage{
setChunkSize(value = 1)
}
\arguments{
\item{value}{The number of iterations to group}
}
\description{
Groups iterations of the foreach loop together per task.
}
\examples{
setChunkSize(10)
}
