% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{setCredentials}
\alias{setCredentials}
\title{Set azure credentials to R session.}
\usage{
setCredentials(fileName = "az_config.json")
}
\arguments{
\item{fileName}{The cluster configuration that was created in \code{makeCluster}}
}
\description{
Set azure credentials to R session.
}

