% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{setCredentialsObject}
\alias{setCredentialsObject}
\title{Set azure credentials to R session from specified R object.}
\usage{
setCredentialsObject(credentialsConfig)
}
\arguments{
\item{credentialsConfig}{The credentials R object was created by user programatically}
}
\description{
Set azure credentials to R session from specified R object.
}
