% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{setHttpTraffic}
\alias{setHttpTraffic}
\title{Set the verbosity for calling httr rest api calls}
\usage{
setHttpTraffic(value = FALSE)
}
\arguments{
\item{value}{Boolean value for turning on and off verbose mode}
}
\description{
Set the verbosity for calling httr rest api calls
}
\examples{
setVerbose(TRUE)
}
