% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{setJobAutoDelete}
\alias{setJobAutoDelete}
\title{Specify whether to delete job/jobresult after asychronous job is completed.}
\usage{
setJobAutoDelete(value = TRUE)
}
\arguments{
\item{value}{boolean of TRUE or FALSE}
}
\description{
Specify whether to delete job/jobresult after asychronous job is completed.
}
\examples{
setJobAutoDelete(FALSE)
}
