% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{setReduce}
\alias{setReduce}
\title{Apply reduce function on a group of iterations of the foreach loop together per task.}
\usage{
setReduce(fun = NULL, ...)
}
\arguments{
\item{fun}{The number of iterations to group}

\item{...}{The arguments needed for the reduction function}
}
\description{
Apply reduce function on a group of iterations of the foreach loop together per task.
}
