% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{stopCluster}
\alias{stopCluster}
\title{Deletes the cluster from your Azure account.}
\usage{
stopCluster(cluster)
}
\arguments{
\item{cluster}{The cluster configuration that was created in \code{makeCluster}}
}
\description{
Deletes the cluster from your Azure account.
}
\examples{
\dontrun{
clusterConfiguration <- makeCluster("cluster_settings.json")
stopCluster(clusterConfiguration)
}
}

