% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{terminateJob}
\alias{terminateJob}
\title{Terminate a job}
\usage{
terminateJob(jobId)
}
\arguments{
\item{jobId}{A job id}
}
\description{
Terminate a job
}
\examples{
\dontrun{
terminateJob("job-001")
}
}
