% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{waitForNodesToComplete}
\alias{waitForNodesToComplete}
\title{Polling method to check status of cluster boot up}
\usage{
waitForNodesToComplete(poolId, timeout = 86400)
}
\arguments{
\item{poolId}{The cluster name to poll for}

\item{timeout}{Timeout in seconds, default timeout is one day}
}
\description{
Polling method to check status of cluster boot up
}
\examples{
\dontrun{
waitForNodesToComplete(clusterId = "testCluster", timeout = 3600)
}
}

