% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-job.R
\name{deleteJob}
\alias{deleteJob}
\title{Delete a job}
\usage{
deleteJob(jobId, verbose = TRUE)
}
\arguments{
\item{jobId}{A job id}
}
\description{
Delete a job
}
\examples{
\dontrun{
deleteJob("job-001")
}
}
