% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{generateClusterConfig}
\alias{generateClusterConfig}
\title{Creates a configuration file for the user's cluster setup.}
\usage{
generateClusterConfig(fileName, ...)
}
\arguments{
\item{fileName}{Cluster settings file name}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates a configuration file for the user's cluster setup.
}
\examples{
{
generateClusterConfig("test_config.json")
generateClusterConfig("test_config.json")
}

}

