% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{generateCredentialsConfig}
\alias{generateCredentialsConfig}
\title{Creates a credentials file for rAzureBatch package authentication}
\usage{
generateCredentialsConfig(fileName, ...)
}
\arguments{
\item{fileName}{Credentials file name}

\item{...}{Further named parameters
\itemize{
 \item{"batchAccount"}: {A list of files that the Batch service will download to the compute node before running the command line.}
 \item{"batchKey"}: {Arguments in the foreach parameters that will be used for the task running.}
 \item{"batchUrl"}: {A list of packages that the Batch service will download to the compute node.}
 \item{"storageAccount"}: {The R environment that the task will run under.}
 \item{"storageKey"}: {The R environment that the task will run under.}
}}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates a credentials file for rAzureBatch package authentication
}
\examples{
{
generateCredentialsConfig("test_config.json")
generateCredentialsConfig("test_config.json", batchAccount = "testbatchaccount",
   batchKey = "test_batch_account_key", batchUrl = "http://testbatchaccount.azure.com",
   storageAccount = "teststorageaccount", storageKey = "test_storage_account_key")
}
}

