% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getJobList}
\alias{getJobList}
\title{Get a list of job statuses from the given job ids}
\usage{
getJobList(jobIds = c())
}
\arguments{
\item{jobIds}{A character vector of job ids}
}
\description{
Get a list of job statuses from the given job ids
}
\examples{
getJobList(c("job-001", "job-002"))
}

