% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobUtilities.R
\name{getJobResult}
\alias{getJobResult}
\title{Download the results of the job}
\usage{
getJobResult(jobId)
}
\arguments{
\item{jobId}{The jobId to download from}
}
\value{
The results from the job.
}
\description{
Download the results of the job
}
\examples{
\dontrun{
getJobResult(jobId = "job-001")
}
}
