% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{makeCluster}
\alias{makeCluster}
\title{Creates an Azure cloud-enabled cluster.}
\usage{
makeCluster(fileName = "az_config.json", fullName = FALSE, wait = TRUE)
}
\arguments{
\item{fileName}{A string that uniquely identifies the job within the account.}

\item{fullName}{A boolean flag for checking the file full name}

\item{wait}{A boolean flag to wait for all nodes to boot up}
}
\value{
The request to the Batch service was successful.
}
\description{
Creates an Azure cloud-enabled cluster.
}
\examples{
cluster <- makeCluster("", fullName = TRUE, wait = TRUE)
}

