% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{setCredentials}
\alias{setCredentials}
\title{Deletes the cluster from your Azure account.}
\usage{
setCredentials(fileName = "az_config.json")
}
\arguments{
\item{fileName}{The cluster configuration that was created in \code{makeCluster}}
}
\value{
The request to the Batch service was successful.
}
\description{
Deletes the cluster from your Azure account.
}
\examples{
clusterConfiguration <- makeCluster("cluster_settings.json")
stopCluster(clusterConfiguration)
}

