% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{setCredentials}
\alias{setCredentials}
\title{Set azure credentials to R session from credentials object or json file.}
\usage{
setCredentials(credentials = "az_config.json", verbose = TRUE)
}
\arguments{
\item{credentials}{The credentials object or json file}
}
\description{
Set azure credentials to R session from credentials object or json file.
}
