% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAzureParallel.R
\name{setReduce}
\alias{setReduce}
\title{Groups iterations of the foreach loop together per task.}
\usage{
setReduce(fun = NULL)
}
\arguments{
\item{fun}{The number of iterations to group}
}
\description{
Groups iterations of the foreach loop together per task.
}
\examples{
setReduce(list)
}

